refactor: app module package
